Norm exec_pipes
