+++ Makefile
@@ -6,7 +6,7 @@
 #    By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/09/03 22:33:20 by irychkov          #+#    #+#              #
-#    Updated: 2024/10/23 16:36:16 by irychkov         ###   ########.fr        #
+#    Updated: 2024/10/23 16:56:29 by irychkov         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -19,7 +19,7 @@ SRCS = src/main.c src/split_args.c src/split_args_helper.c \
 		src/parse_redirect.c src/parse_redirect2.c \
 		src/check_syntax.c src/check_syntax2.c src/check_syntax3.c\
 		src/parse_trim_spaces.c \
-		src/exec_child.c \
+		src/exec_child.c src/exec_in_parent.c \
 		src/exec_cmd.c src/exec_pipes.c \
 		src/errors.c src/sys_errors.c\
 		src/free.c \
